requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
spacy==3.7.4
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
